feat: added casting methid to convert the data types

+++ ImplicitConversionPractice/Program.cs
@@ -7,7 +7,7 @@ namespace ImplicitConversionPractice
         public static void Main(string[] args)
         {
             float myFloatValue = 10;
-            int myIntValue = myFloatValue;
+            int myIntValue = (int) myFloatValue;
             Console.WriteLine("myIntValue is " + myIntValue);
         }
     }
